numpy "b" types are signed 8bits integers, a type not available in DAP2. Promoting these to signed 16bits integers is the only options

+++ src/pydap/responses/dds.py
@@ -29,7 +29,7 @@ typemap = {
     'f': 'Float32',
     'h': 'Int16',
     'i': 'Int32', 'l': 'Int32', 'q': 'Int32',
-    'b': 'Byte',
+    'b': 'Int16',
     'H': 'UInt16',
     'I': 'UInt32', 'L': 'UInt32', 'Q': 'UInt32',
     'B': 'Byte',
